move viem-based isCel2() to a separate file
